+++ src/CoreEx.Cosmos/CosmosDbArgs.cs
@@ -1,5 +1,6 @@
 ﻿// Copyright (c) Avanade. Licensed under the MIT License. See https://github.com/Avanade/CoreEx
 
+using CoreEx.Entities;
 using Microsoft.Azure.Cosmos;
 using System;
 using System.Net;
@@ -30,6 +31,8 @@ namespace CoreEx.Cosmos
             CleanUpResult = template.CleanUpResult;
             FilterByTenantId = template.FilterByTenantId;
             GetTenantId = template.GetTenantId;
+            FormatIdentifier = template.FormatIdentifier;
+            ParseIdentifier = template.ParseIdentifier;
         }
 
         /// <summary>
@@ -94,5 +97,51 @@ namespace CoreEx.Cosmos
         /// Gets or sets the <i>get</i> tenant identifier function; defaults to <see cref="ExecutionContext.Current"/> <see cref="ExecutionContext.TenantId"/>.
         /// </summary>
         public Func<string?> GetTenantId { get; set; } = () => ExecutionContext.HasCurrent ? ExecutionContext.Current.TenantId : null;
+
+        /// <summary>
+        /// Formats an identifier to a <see cref="string"/> representation based on its underlying <see cref="Type"/> (used by <see cref="CosmosDbContainerBase{T, TModel, TSelf}.GetCosmosId(T)"/> and <see cref="ICosmosDbValue.PrepareBefore"/>).
+        /// </summary>
+        /// <returns>The identifier as a <see cref="string"/>.</returns>
+        public Func<object?, string> FormatIdentifier { get; set; } = DefaultFormatIdentifier;
+
+        /// <summary>
+        /// Parses a <see cref="string"/> identifier representation into its underlying <see cref="Type"/> (used by the <see cref="ICosmosDbValue.PrepareAfter"/>).
+        /// </summary>
+        /// <returns>The parsed identifier.</returns>
+        public Action<object, string?> ParseIdentifier { get; set; } = DefaultParseIdentifier;
+
+        /// <summary>
+        /// Provides the default <see cref="FormatIdentifier"/> implementation.
+        /// </summary>
+        public static Func<object?, string> DefaultFormatIdentifier { get; } = id => id.ThrowIfNull(nameof(id)) switch
+        {
+            string s => s ?? throw new InvalidOperationException("The identifier must not be null."),
+            int i => i.ToString(System.Globalization.CultureInfo.InvariantCulture),
+            long l => l.ToString(System.Globalization.CultureInfo.InvariantCulture),
+            Guid g => g.ToString(),
+            CompositeKey ck => ck.ToString() ?? throw new InvalidOperationException("The identifier must not be null."),
+            _ => throw new NotSupportedException("An identifier must be one of the following types: string, int, long, Guid or CompositeKey.")
+        };
+
+        /// <summary>
+        /// Provides the default <see cref="ParseIdentifier"/> implementation.
+        /// </summary>
+        public static Action<object, string?> DefaultParseIdentifier { get; } = (value, id) => 
+        {
+            value.ThrowIfNull(nameof(value));
+            if (value is IIdentifier iid)
+            {
+                iid.Id = iid.IdType switch
+                {
+                    Type t when t == typeof(string) => id,
+                    Type t when t == typeof(int) => id == null ? 0 : int.Parse(id, System.Globalization.CultureInfo.InvariantCulture),
+                    Type t when t == typeof(long) => id == null ? 0 : long.Parse(id, System.Globalization.CultureInfo.InvariantCulture),
+                    Type t when t == typeof(Guid) => id == null ? Guid.Empty : Guid.Parse(id),
+                    _ => throw new NotSupportedException("An IIdentifier.IdType must be one of the following types: string, int, long, or Guid.")
+                };
+
+                return;
+            }
+        };
     }
 }