+++ src/CoreEx.Cosmos/CosmosDbModelQueryBase.cs
@@ -15,26 +15,17 @@ namespace CoreEx.Cosmos
     /// </summary>
     /// <typeparam name="TModel">The cosmos model <see cref="Type"/>.</typeparam>
     /// <typeparam name="TSelf">The <see cref="Type"/> itself.</typeparam>
-    public abstract class CosmosDbModelQueryBase<TModel, TSelf> where TModel : new() where TSelf : CosmosDbModelQueryBase<TModel, TSelf>
+    public abstract class CosmosDbModelQueryBase<TModel, TSelf>(ICosmosDbContainer container, CosmosDbArgs dbArgs) where TModel : new() where TSelf : CosmosDbModelQueryBase<TModel, TSelf>
     {
-        /// <summary>
-        /// Initializes a new instance of the <see cref="CosmosDbModelQueryBase{TModel, TSelf}"/> class.
-        /// </summary>
-        protected CosmosDbModelQueryBase(ICosmosDbContainer container, CosmosDbArgs dbArgs)
-        {
-            Container = container.ThrowIfNull(nameof(container));
-            QueryArgs = dbArgs;
-        }
-
         /// <summary>
         /// Gets the <see cref="ICosmosDbContainer"/>.
         /// </summary>
-        public ICosmosDbContainer Container { get; }
+        public ICosmosDbContainer Container { get; } = container.ThrowIfNull(nameof(container));
 
         /// <summary>
         /// Gets the <see cref="CosmosDbArgs"/>.
         /// </summary>
-        public CosmosDbArgs QueryArgs;
+        public CosmosDbArgs QueryArgs = dbArgs;
 
         /// <summary>
         /// Gets the <see cref="PagingResult"/>.