+++ src/CoreEx.Cosmos/CosmosDbQuery.cs
@@ -18,7 +18,7 @@ namespace CoreEx.Cosmos
     /// <param name="container">The <see cref="CosmosDbContainer{T, TModel}"/>.</param>
     /// <param name="dbArgs">The <see cref="CosmosDbArgs"/>.</param>
     /// <param name="query">A function to modify the underlying <see cref="IQueryable{T}"/>.</param>
-    public class CosmosDbQuery<T, TModel>(CosmosDbContainer<T, TModel> container, CosmosDbArgs dbArgs, Func<IQueryable<TModel>, IQueryable<TModel>>? query) : CosmosDbQueryBase<T, TModel, CosmosDbQuery<T, TModel>>(container, dbArgs) where T : class, IEntityKey, new() where TModel : class, IIdentifier<string>, new()
+    public class CosmosDbQuery<T, TModel>(CosmosDbContainer<T, TModel> container, CosmosDbArgs dbArgs, Func<IQueryable<TModel>, IQueryable<TModel>>? query) : CosmosDbQueryBase<T, TModel, CosmosDbQuery<T, TModel>>(container, dbArgs) where T : class, IEntityKey, new() where TModel : class, IEntityKey, new()
     {
         private readonly Func<IQueryable<TModel>, IQueryable<TModel>>? _query = query;