+++ src/CoreEx.Cosmos/ICosmosDb.cs
@@ -49,7 +49,7 @@ namespace CoreEx.Cosmos
         /// <param name="containerId">The <see cref="Container"/> identifier.</param>
         /// <param name="dbArgs">The <see cref="CosmosDbArgs"/>.</param>
         /// <returns>The <see cref="CosmosDbContainer{T, TModel}"/>.</returns>
-        CosmosDbContainer<T, TModel> Container<T, TModel>(string containerId, CosmosDbArgs? dbArgs = null) where T : class, IEntityKey, new() where TModel : class, IIdentifier<string>, new();
+        CosmosDbContainer<T, TModel> Container<T, TModel>(string containerId, CosmosDbArgs? dbArgs = null) where T : class, IEntityKey, new() where TModel : class, IEntityKey, new();
 
         /// <summary>
         /// Gets (creates) the <see cref="CosmosDbValueContainer{T, TModel}"/> for the specified <paramref name="containerId"/>.
@@ -59,7 +59,7 @@ namespace CoreEx.Cosmos
         /// <param name="containerId">The <see cref="Container"/> identifier.</param>
         /// <param name="dbArgs">The <see cref="CosmosDbArgs"/>.</param>
         /// <returns>The <see cref="CosmosDbValueContainer{T, TModel}"/>.</returns>
-        CosmosDbValueContainer<T, TModel> ValueContainer<T, TModel>(string containerId, CosmosDbArgs? dbArgs = null) where T : class, IEntityKey, new() where TModel : class, IIdentifier, new();
+        CosmosDbValueContainer<T, TModel> ValueContainer<T, TModel>(string containerId, CosmosDbArgs? dbArgs = null) where T : class, IEntityKey, new() where TModel : class, IEntityKey, new();
 
         /// <summary>
         /// Gets (creates) the <see cref="CosmosDbModelQuery{TModel}"/> for the specified <paramref name="containerId"/>.
@@ -69,7 +69,7 @@ namespace CoreEx.Cosmos
         /// <param name="dbArgs">The <see cref="CosmosDbArgs"/>.</param>
         /// <param name="query">A function to modify the underlying <see cref="IQueryable{T}"/>.</param>
         /// <returns>The <see cref="CosmosDbModelQuery{TModel}"/>.</returns>
-        CosmosDbModelQuery<TModel> ModelQuery<TModel>(string containerId, CosmosDbArgs dbArgs, Func<IQueryable<TModel>, IQueryable<TModel>>? query) where TModel : class, IIdentifier<string>, new();
+        CosmosDbModelQuery<TModel> ModelQuery<TModel>(string containerId, CosmosDbArgs dbArgs, Func<IQueryable<TModel>, IQueryable<TModel>>? query) where TModel : class, IEntityKey, new();
 
         /// <summary>
         /// Gets (creates) the <see cref="CosmosDbValueModelQuery{TModel}"/> for the specified <paramref name="containerId"/>.
@@ -79,7 +79,7 @@ namespace CoreEx.Cosmos
         /// <param name="dbArgs">The <see cref="CosmosDbArgs"/>.</param>
         /// <param name="query">A function to modify the underlying <see cref="IQueryable{T}"/>.</param>
         /// <returns>The <see cref="CosmosDbValueModelQuery{TModel}"/>.</returns>
-        CosmosDbValueModelQuery<TModel> ValueModelQuery<TModel>(string containerId, CosmosDbArgs dbArgs, Func<IQueryable<CosmosDbValue<TModel>>, IQueryable<CosmosDbValue<TModel>>>? query) where TModel : class, IIdentifier<string>, new();
+        CosmosDbValueModelQuery<TModel> ValueModelQuery<TModel>(string containerId, CosmosDbArgs dbArgs, Func<IQueryable<CosmosDbValue<TModel>>, IQueryable<CosmosDbValue<TModel>>>? query) where TModel : class, IEntityKey, new();
 
         /// <summary>
         /// Invoked where a <see cref="CosmosException"/> has been thrown.
@@ -97,7 +97,7 @@ namespace CoreEx.Cosmos
         /// <typeparam name="TModel">The cosmos model <see cref="Type"/>.</typeparam>
         /// <param name="dbArgs">The <see cref="CosmosDbArgs"/>.</param>
         /// <returns>The <see cref="ItemRequestOptions"/>.</returns>
-        ItemRequestOptions GetItemRequestOptions<T, TModel>(CosmosDbArgs dbArgs) where T : class, new() where TModel : class, IIdentifier, new();
+        ItemRequestOptions GetItemRequestOptions<T, TModel>(CosmosDbArgs dbArgs) where T : class, new() where TModel : class, IEntityKey, new();
 
         /// <summary>
         /// Gets or instantiates the <see cref="QueryRequestOptions"/>.
@@ -106,7 +106,7 @@ namespace CoreEx.Cosmos
         /// <typeparam name="TModel">The cosmos model <see cref="Type"/>.</typeparam>
         /// <param name="dbArgs">The <see cref="CosmosDbArgs"/>.</param>
         /// <returns>The <see cref="QueryRequestOptions"/>.</returns>
-        QueryRequestOptions GetQueryRequestOptions<T, TModel>(CosmosDbArgs dbArgs) where T : class, new() where TModel : class, IIdentifier, new();
+        QueryRequestOptions GetQueryRequestOptions<T, TModel>(CosmosDbArgs dbArgs) where T : class, new() where TModel : class, IEntityKey, new();
 
         /// <summary>
         /// Gets or instantiates the <see cref="QueryRequestOptions"/>.
@@ -114,7 +114,7 @@ namespace CoreEx.Cosmos
         /// <typeparam name="TModel">The cosmos model <see cref="Type"/>.</typeparam>
         /// <param name="dbArgs">The <see cref="CosmosDbArgs"/>.</param>
         /// <returns>The <see cref="QueryRequestOptions"/>.</returns>
-        QueryRequestOptions GetQueryRequestOptions<TModel>(CosmosDbArgs dbArgs) where TModel : class, IIdentifier, new();
+        QueryRequestOptions GetQueryRequestOptions<TModel>(CosmosDbArgs dbArgs) where TModel : class, IEntityKey, new();
 
         /// <summary>
         /// Gets the authorization filter.
@@ -123,20 +123,5 @@ namespace CoreEx.Cosmos
         /// <param name="containerId">The <see cref="Microsoft.Azure.Cosmos.Container"/> identifier.</param>
         /// <returns>The filter query where found; otherwise, <c>null</c>.</returns>
         Func<IQueryable, IQueryable>? GetAuthorizeFilter<TModel>(string containerId);
-
-        /// <summary>
-        /// Formats an identifier to a <see cref="string"/> representation based on its underlying <see cref="Type"/> (used by <see cref="CosmosDbContainerBase{T, TModel, TSelf}.GetCosmosId(T)"/> and <see cref="ICosmosDbValue.PrepareBefore"/>).
-        /// </summary>
-        /// <param name="id">The identifier.</param>
-        /// <returns>The identifier as a <see cref="string"/>.</returns>
-        string FormatIdentifier(object? id);
-
-        /// <summary>
-        /// Parses a <see cref="string"/> identifier representation into its underlying <see cref="Type"/> (used by the <see cref="ICosmosDbValue.PrepareAfter"/>).
-        /// </summary>
-        /// <param name="type">The identifier <see cref="Type"/>.</param>
-        /// <param name="id">The identifier as a <see cref="string"/>.</param>
-        /// <returns>The parsed identifier.</returns>
-        object? ParseIdentifier(Type type, string? id);
     }
 }