+++ src/CoreEx.Cosmos/ICosmosDbContainerT.cs
@@ -14,7 +14,7 @@ namespace CoreEx.Cosmos
     /// </summary>
     /// <typeparam name="T">The entity <see cref="Type"/>.</typeparam>
     /// <typeparam name="TModel">The cosmos model <see cref="Type"/>.</typeparam>
-    public interface ICosmosDbContainer<T, TModel> : ICosmosDbContainer where T : class, IEntityKey, new() where TModel : class, IIdentifier, new()
+    public interface ICosmosDbContainer<T, TModel> : ICosmosDbContainer where T : class, IEntityKey, new() where TModel : class, IEntityKey, new()
     {
         /// <summary>
         /// Gets the <see cref="PartitionKey"/>.